refactor(core): split temporal preflight fields from annotations
Extract preflightTemporalFields (mapped column parse/inference) and
preflightTemporalAnnotations (rowless intercept checks) from the 439-line
preflightTemporalBindings orchestrator. Shared assertTemporalConfiguration
and docs live in temporal-preflight-shared; facade preserves public import
path and phase order (fields → commonConversion → annotations).
