+++ .changeset/core-temporal-preflight-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/core": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+refactor: split temporal preflight into field, annotation, and shared modules
